+++ Packages/com.unity.cloud.gltfast/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - (Tests) Reduced jobs performance test duration by lowering buffer sizes and switching to dynamic measurement counts.
+- (Tests) Performance tests are not run, unless the `RUN_PERFORMANCE_TESTS` scripting define is set.
 
 ### Fixed
 - Performance test compilation if *Collections* package >= 1.5.0 is not installed.